One more bugfix

+++ showhdr.c
@@ -5,12 +5,12 @@
 
 #define GET_IMAGE_TITLE(X) ((uint8_t *)X) /* 32? */
 #define BIOS_INFO_GET_LOGICAL_AREA_TYPE(X) (*((uint8_t *)X+32))
-#define BIOS_INFO_GET_LOGICAL_AREA_SIZE(X) (*(uint16_t *)(((uint8_t *)X)+33))
+#define BIOS_INFO_GET_LOGICAL_AREA_SIZE(X) (*(uint32_t *)(((uint8_t *)X)+33))
 #define BIOS_INFO_GET_LOAD_FROM_FILE(X) (*((uint8_t *)X+37))
 #define BIOS_INFO_GET_REBOOT_REQUIRED(X) (*((uint8_t *)X+38))
 #define BIOS_INFO_GET_UPDATE_ALL_OF_IMAGE(X) (*((uint8_t *)X+39))
-#define BIOS_INFO_GET_LOGICAL_AREA_STRING(X) ((uint8_t *)X+40) /* 16? */
-#define GET_TIME_STAMP(X) ((uint8_t *)X+64) /* 19? */
+#define BIOS_INFO_GET_LOGICAL_AREA_STRING(X) ((uint8_t *)X+40) /* 24? */
+#define GET_TIME_STAMP(X) ((uint8_t *)X+64) /* 16? */
 #define GET_THIS_FILE_START_ADDR(X) (*(uint32_t *)(((uint8_t *)X)+80))
 #define GET_THIS_FILE_DATA_LENGTH(X) (*(uint32_t *)(((uint8_t *)X)+84))
 #define GET_LOGICAL_AREA_TYPE(X) (*((uint8_t *)X+88))
